requirements.txt
Django~=4.2.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
numpy==1.24.3
Pillow==9.3.0
requests==2.27.1
python-dotenv~=0.21.0
django-filter==23.2
pylint==2.17.4
autopep8~=2.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-redis==5.2.0
django-environ==0.7.0
django-taggit==4.0.0
